pyproject.toml
[tool.poetry]
name = "habitsproductivity"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

packages = [{ include = "habits" }]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
celery = "^5.5.3"
redis = "^6.2.0"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.10"
git-filter-repo = "^2.47.0"
swagger-ui-bundle = "^0.0.9"
# redoc = "^1.2.0"
djangorestframework-simplejwt = "^5.5.0"
django-celery-beat = "^2.8.1"
python-telegram-bot = "^22.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
coverage = "^7.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
